Sửa lỗi tăng giảm số lượng cho phép chọn edit text

+++ LibMagestore/src/main/java/com/magestore/app/view/EditTextFloat.java
@@ -80,7 +80,7 @@ public class EditTextFloat extends EditText {
         return ConfigUtil.parseDouble(super.getText().toString());
     }
 
-    public float getValueInteger() {
+    public int getValueInteger() {
         return ConfigUtil.parseInteger(super.getText().toString());
     }
